+++ Editor/Scripts/UIAutoReference.cs
@@ -11,7 +11,7 @@ using Feif.Extensions;
 
 namespace Feif.UIFramework.Editor
 {
-    public class UIAutoReference
+    public static class UIAutoReference
     {
         [InitializeOnLoadMethod]
         public static void AddListeners()
@@ -53,6 +53,10 @@ namespace Feif.UIFramework.Editor
             }
         }
 
+        /// <summary>
+        /// 自动引用，并维护UI关系树
+        /// </summary>
+        /// <returns>被自动引用赋值的的字段</returns>
         public static List<string> SetReference(UnityObject script)
         {
             var result = new List<string>();
@@ -116,6 +120,27 @@ namespace Feif.UIFramework.Editor
                 field.SetValue(script, isGameObject ? target.gameObject as UnityObject : target);
             }
             EditorUtility.SetDirty(script);
+            return result;
+        }
+
+        /// <summary>
+        /// 扫描目录下所有Prefab，并自动引用UIBase组件中的值
+        /// </summary>
+        /// <param name="path">要扫描的目录，例如：Assets/Prefabs</param>
+        /// <returns>被自动引用的资源(资源路径)</returns>
+        public static List<string> SetReference(string path)
+        {
+            var result = new List<string>();
+            var paths = AssetDatabase.GetAllAssetPaths().Where(item => item.StartsWith(path) && item.EndsWith(".prefab"));
+            foreach (var item in paths)
+            {
+                var asset = AssetDatabase.LoadAssetAtPath<GameObject>(item);
+                if (asset == null) continue;
+                var uibase = asset.GetComponent<UIBase>();
+                if (uibase == null) continue;
+                SetReference(uibase);
+                result.Add(item);
+            }
             return result;
         }
     }