+++ Editor/Scripts/UIFrameSetting.cs
@@ -1,3 +1,5 @@
+using System.IO;
+using UnityEditor;
 using UnityEngine;
 
 namespace Feif.UIFramework.Editor
@@ -10,5 +12,23 @@ namespace Feif.UIFramework.Editor
         public TextAsset UIPanelTemplate;
         public TextAsset UIWindowTemplate;
         public bool AutoReference = true;
+
+        private void Reset()
+        {
+            var ms = MonoScript.FromScriptableObject(this);
+            var path = AssetDatabase.GetAssetPath(ms);
+            var resPath = Path.GetDirectoryName(path).Replace("Scripts", "Resources");
+            var fields = GetType().GetFields();
+            foreach (var field in fields)
+            {
+                if (field.Name.EndsWith("Template"))
+                {
+                    var file = Path.Combine(resPath, $"{field.Name}.txt");
+                    var res = AssetDatabase.LoadAssetAtPath<TextAsset>(file);
+                    field.SetValue(this, res);
+                }
+            }
+            EditorUtility.SetDirty(this);
+        }
     }
 }